+++ HISTORY.rst
@@ -2,6 +2,13 @@
 History
 =======
 
+2.0.2 (2017-05-18)
+------------------
+
+    * Handle ValidationError on collection_post and put  (ericof).
+    * Change newrelic transaction name to use instance class and module name (rudaporto).
+    * Fix: bad check condition was avoiding request to be appended in the model instance (rudaporto).
+
 2.0.1 (2017-05-03)
 ------------------
     * Lookup for _default_notify_events on the model before look in the resource view (rudaporto).